+++ R/L1centNB.R
@@ -9,10 +9,10 @@
 #' identical.
 #'
 #' @details
-#' For an undirected graph, if the graph is symmetrized (in a way defined in Kang and Oh 2025a)
+#' For an undirected graph, if the graph is symmetrized (in a way defined in Kang and Oh 2026a)
 #' w.r.t. a vertex \ifelse{html}{\out{<i>v</i>}}{\eqn{v}}, vertex
 #' \ifelse{html}{\out{<i>v</i>}}{\eqn{v}} becomes the graph median (Kang and Oh
-#' 2025a), i.e., \ifelse{html}{\out{<i>v</i>}}{\eqn{v}} has
+#' 2026a), i.e., \ifelse{html}{\out{<i>v</i>}}{\eqn{v}} has
 #' \ifelse{html}{\out{<i>L</i><sub>1</sub>}}{{\eqn{L_1}}} centrality 1. Based on
 #' this property, we define the
 #' \ifelse{html}{\out{<i>L</i><sub>1</sub>}}{{\eqn{L_1}}} centrality-based
@@ -23,7 +23,7 @@
 #'
 #' For a directed graph, a vertex of interest, say
 #' \ifelse{html}{\out{<i>v</i>}}{\eqn{v}}, is made to a centrality and prestige
-#' median vertex by the procedure described in Kang and Oh (2025b). We call the
+#' median vertex by the procedure described in Kang and Oh (2026b). We call the
 #' resulting graph as the modified graph w.r.t.
 #' \ifelse{html}{\out{<i>v</i>}}{\eqn{v}}.
 #' \ifelse{html}{\out{<i>L</i><sub>1</sub>}}{{\eqn{L_1}}}
@@ -62,36 +62,43 @@
 #'   [L1centLOC()] and [L1centEDGE()] internally uses \code{L1centNB()}.
 #
 #' @examples
-#' NB <- L1centNB(MCUmovie, eta = igraph::V(MCUmovie)$worldwidegross)
+#' NB <- L1centNB(MCUmovie, vertex_weight = igraph::V(MCUmovie)$worldwidegross)
 #' # Top 6 L1 centrality-based neighbors of "Iron Man"
 #' utils::head(sort(NB$"Iron Man", decreasing = TRUE))
 #' @references S. Kang and H.-S. Oh. On a notion of graph centrality based on
 #'   \ifelse{html}{\out{<i>L</i><sub>1</sub>}}{{\eqn{L_1}}} data depth.
-#'   \emph{Journal of the American Statistical Association}, 1--13, 2025a.
+#'   \emph{Journal of the American Statistical Association}, 121(553): 400--412, 2026a.
 #'
 #'   S. Kang and H.-S. Oh.
 #'   \ifelse{html}{\out{<i>L</i><sub>1</sub>}}{{\eqn{L_1}}} prominence measures
-#'   for directed graphs. \emph{The American Statistician}, 1--16, 2025b.
-L1centNB <- function(g, eta, mode, weight_transform) UseMethod("L1centNB")
+#'   for directed graphs. \emph{The American Statistician}, 80(2): 301--309, 2026b.
+L1centNB <- function(g, vertex_weight, mode, edge_weight_transform) UseMethod("L1centNB")
 
 #' @name L1centNB
 #' @exportS3Method L1centNB igraph
-L1centNB.igraph <- function(g, eta = NULL, mode = c("centrality", "prestige"), weight_transform = NULL){
+L1centNB.igraph <- function(g, vertex_weight = NULL, mode = c("centrality", "prestige"), edge_weight_transform = NULL){
   validate_igraph(g, checkdir = FALSE)
   mode <- match.arg(tolower(mode), choices = c("centrality", "prestige"))
 
-  new_weight <- edge_weight_transform(g, weight_transform)
+  new_weight <- edge_weight_transform0(g, edge_weight_transform)
   if(!is.null(new_weight)) igraph::E(g)$weight <- new_weight
   D <- igraph::distances(g, mode = "out")
-  L1centNB.matrix(D, eta, mode = mode)
+  L1centNB.matrix(D, vertex_weight, mode = mode)
 }
 
 #' @name L1centNB
 #' @exportS3Method L1centNB matrix
-L1centNB.matrix <- function(g, eta = NULL, mode = c("centrality", "prestige"), weight_transform = NULL){
+L1centNB.matrix <- function(g, vertex_weight = NULL, mode = c("centrality", "prestige"), edge_weight_transform = NULL){
+  eta <- vertex_weight
   if(is.null(eta)) eta <- rep(1,ncol(g))
   validate_matrix(g, eta, checkdir = FALSE)
   mode <- match.arg(tolower(mode), choices = c("centrality", "prestige"))
+
+  calls <- sys.calls()
+  fnames <- sapply(calls, function(cl) as.character(cl[[1]]))
+  # print(fnames)
+  if(all(fnames[1:2] == c("L1centNB", "L1centNB.matrix")))
+    message("DISTANCE matrix received")
 
   n <- ncol(g)
   sg <- ifelse(isSymmetric.matrix(g), 1,
@@ -101,7 +108,7 @@ L1centNB.matrix <- function(g, eta = NULL, mode = c("centrality", "prestige"), w
   NB <- lapply(1:n, function(i){
     neweta <- eta
     neweta[i] <- etasum/sg + neweta[i]
-    c(L1cent(g,eta=neweta,mode=mode))
+    c(L1cent(g,vertex_weight=neweta,mode=mode))
   })
   names(NB) <- label
   structure(NB,